+++ docs/CHANGELOG.md
@@ -11,6 +11,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   definitions (OSIDB-4938)
 - Added `bu_labels` label type (OSIDB-4982)
 
+### Changed
+- Allow adding labels to flaws in any workflow state, including DONE (OSIDB-4986)
+
+### Fixed
+- Exclude streams with no offer from tracker suggestions response (OSIDB-4977)
+
 ## [5.10.3] - 2026-05-06
 ### Fixed
 - Include builds during affect auto-creation (OSIDB-4965)